// src/screens/auth/signup/recruiterDetails/types.ts
export type IStateRecruiterProps = {
  name: string;
  email: string;
  company: string;
  phoneNumber: string;
  location: string;
  locationError: string;
  phoneNumberError: string;
  nameError: string;
  industry: string;
  industryError: string;
  companyError: string;
};

export type IRecruiterProps = {};